+++ src/main/java/org/jnjeaaaat/snms/global/exception/GlobalExceptionHandler.java
@@ -1,8 +1,8 @@
-package org.jnjeaaaat.snms.global.auth.exception;
+package org.jnjeaaaat.snms.global.exception;
 
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
-import org.jnjeaaaat.snms.global.auth.exception.dto.ErrorResponse;
+import org.jnjeaaaat.snms.global.exception.dto.ErrorResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
